requirements.txt
Flask
Pillow
numpy
cryptography
